2838 - it should show that at the top of the screen updated command line.. run -help for some parameters for loop options, new -debug flag change that now keeps the studio unless you check it get it from imdb again during autopilot (at djh_ and digital's request)

+++ movieinfoplus/modNfoData.vb
@@ -31,14 +31,14 @@ Module modNfoData
             If tmovie.pimdbnumber = "" Then
                 Return False
             Else
-                File.Move(pathtonfo, pathtonfo & ".oldversion")
+                If Not File.Exists(pathtonfo & ".oldversion") Then File.Move(pathtonfo, pathtonfo & ".oldversion")
                 Return True
             End If
             ' Return True
         End If
     End Function
     Public Sub checknfodata(ByRef currentmovie As movie, ByVal dname As String, ByVal workonline As Boolean)
-
+        If currentmovie.pdatafromnfo Then Exit Sub 'nothing to do if it came from an nfo file
         ' ---- IMDB AND NFO -----
         Dim hasnfoalready As Boolean = False
         Dim haveidonly As Boolean = False
